pyproject.toml
[project]
name = "pullpeek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.74.3",
    "pydantic>=2.11.7",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
]
